requirements.txt
functions-framework==3.*
requests==2.*
beautifulsoup4==4.*
pandas==2.*
pyarrow
google-cloud-storage
google-cloud-bigquery